Bug #207: Improve exception safety with smart pointers - source file changes
The class "unique_ptr" which became officially available with the standard
"ISO/IEC 14882:2011" can help to make the software a bit more robust for the
case that an exception will be eventually thrown. It was introduced therefore
into some source files.

Some explicit calls of the C++ delete operator and unnecessary checks could be
removed because of its reuse.

Signed-off-by: Markus Elfring <elfring@users.sourceforge.net>